Upgrade dependencies
This upgrades dev and prod dependencies with the following commands:

```
just upgrade dev
just upgrade prod
```

I'm unclear why Dependabot missed these, but know that we've had issues
with Dependabot in the past, especially concerning transitive
dependencies (i.e. dependencies of dependencies). Whilst it's not
normally a good idea to upgrade all dependencies in one go,
actions-registry is such a small project that we can mitigate the risk
of something going wrong by running the tests and by checking the
development server for errors. You'll be pleased to learn that the tests
pass and the development server doesn't report any errors.