Order versions from latest to oldest
- We will be making a change to show all past versions of an action to the user.
- To display the action versions from latest to oldest,
a sort is needed either database-side or application-side.
- As of current plans, we would only retrieve versions from the
database to get the latest version or display them in reverse-
chronological order.
- Given that a retrieval is always coupled
to a sort, we can neatly specify the ordering in the `Meta`
class of the `Version` model.
- A corresponding database migration is done, and a unit test
is added to ensure that the commit timestamp is used for ordering.