Use template tag to correctly resolve relative Github links
- We get the readme for an action version as html from Github, and
show this in version.html.
- The readme might link to files in the repo (e.g. images / markdown
files). These are relative links that github does not resolve to
absolute links before passing us the readme html. As a result the
links are broken on our site (and images do not show).
- Use a template tag (herenamed `resolve_github_links` to convert
relative github links to absolute ones.
- `<a>` tags and `<img>` tags are treated separately as the `src`
attribute of the `<img>` tag requires a link to the raw image,
whereas <a> tags should simply send the user to a github site when
clicked.
- The version tag is used in the url templates to ensure that the
links match the readme text being shown.