Add template tag to correctly resolve relative Github links
- We get the readme for an action version as html from Github, and
show this in version.html.

- The readme might link to files in the repo (e.g. images / markdown
files). These are relative links that github does not resolve to
absolute links before passing us the readme html. As a result the
links are broken on our site.

- Use a template tag (herenamed `resolve_relative_links` to convert
relative github links to absolute ones.

- `<a>` tags and `<img>` tags are treated separately - the `src`
attribute of the `<img>` tag requires raw image urls, while we do
not want to pass raw file urls for links that users would click.

- Note that the version tag is used in the url templates to ensure
that the links match the readme text being shown.