Use template tag to correctly resolve relative Github image links
- We get the readme for an action version as html from Github, and
show this in version.html.
- The readme might link to images in the repo. These are relative
links that github does not resolve to absolute links before passing
us the readme html. As a result the image links are broken on our
site.
- To show the images in the actions registry as well, use a template
tag (herenamed `resolve_image_links` to convert relative github
links to absolute ones.
- The version tag is used in the url template to ensure that the
image shown is the one that matches the readme text being shown.