Add utils function to resolve relative urls
- We get the readme for an action version as html from Github, and
show this in version.html.

- The readme might link to files in the repo (e.g. images / markdown
files). These are relative links that github does not resolve to
absolute links before passing us the readme html. As a result the
links are broken on our site.

-  Create a `utils` module. Within it, add a utils function (herenamed
`resolve_relative_urls_to_absolute` to convert relative github links
to absolute ones. We will use this function in a later commit.

- Since we use the `BeautifulSoup` html parsing library,
we add it as a prod dependency.