Resolve relative github urls in `version` view
- In the `version` view, pass `readme` as a separate context item
to the template; the `version.html` template is altered accordingly.
This allows us to resolve relative github urls to absolute ones in
the readme html within the view prior to passing it to
the template.

- `<a>` tags and `<img>` tags are treated separately - the `src`
attribute of the `<img>` tag requires raw image urls, while we do
not want to pass raw file urls for links that users would click.

- Note that the version tag is used in the url templates to ensure
that the links match the readme text being shown.