Remove django-extensions dependency
Now that we have converted the fetch action job to a management command
(6177ef9), we no longer need django-extensions. For more information,
see #409. To get a sense of what django-extensions adds, but we don't
need, run `./manage.py` at this commit and at the previous commit, and
compare the output.