Add a blurb to the homepage in the same div as the table
- Add a blurb that tells a visitor what the site is for, with a link to the
OpenSAFELY site for users who are hearing about OpenSAFELY for the first
time.

- The blurb is placed into a separarate bordered div to the table,
with a gap in between the two elements. The tailwind classes of the
divs are near-identical, with the only difference in styling being that
the blurb has a white background.

- The blurb's horizontal padding (px-6) matches that of the table contents.
The vertical padding (py-8) and the prose CSS matches what is used for the
readme in `version.html` to maintain consistency across the site. Since we
do not include code in the blurb, the `prose-pre` and `prose-code` stylings
are not copied over.