Set `.anchor` elements to `display: none;`
- The `.anchor` elements in our displayed readmes (including its
children svg elements) lead to large spaces under headings. The
link is hidden somehwere within this large gap, and upon clicking
sends the user to the readme on github. This looks awkward and is
unhelpful.

- We set these `.anchor` elements to `display: none;`, so that they
are removed from the page and no longer leave a large white gap
under the headings.

- We also remove the styling that applies to the now removed elements,
since they no longer have an effect on the page.