Fetch human contributors to repo in `fetch_action`
- In the `fetch_action` management command,
fetch all contributors to the repo but filter out bots,
since in actions-registry we only want to show human
contributors.

- Helpfully, the github API endpoint tags bot contributors with
`[bot]` in their login, so this is used to filter them out.
Note that we do not want to filter out usernames that happen
to contain the string "bot" for unrelated reasons.