Fetch repo topics in `fetch_action`
- Fetch topics from GitHub and ensure that they are ordered alphabetically.

- Calling both `client.get_repo()` and `repo.get_topics()` without
using a cache means that the repo endpoint would be called twice
per repo when running the management command.
See opensafely-core/osgithub#221 for details.

- The number of duplicate calls is equal to `Action.objects.count()`,
which is at the moment equal to 10. Even including these duplicate calls,
we are not close to hitting GitHub's API rate limit of 5000 calls,
so we opt to make the duplicate calls as opposed to caching the requests,
which adds complexity to the codebase.