Add django-debug-toolbar as a dev dependency
- We would like to use the toolbar for debugging purposes.
Add it via `just requirements-dev`.

- Make changes to `settings.py` and `urls.py` based on django debug
toolbar's installation docs[^1].
Some of these changes cause an error if django-debug-toolbar
is unavailable, which is the case for our production environment.

- Hence, make the changes conditional on a boolean environment variable
named `DJANGO_DEBUG_TOOLBAR`. We define this new environment variable
instead of reusing `DEBUG` - this is to keep the behavioural changes
from the toolbar untangled from those of django's debug mode, so that
there are fewer moving parts during debugging.

- Since we would like to use the django debug toolbar when
and only when we run the local development server,
set `DJANGO_DEBUG_TOOLBAR` to 1 in the `just run` recipe but not elsewhere.

- Note that while we add `INTERNAL_IPS` in `settings.py`` since the
toolbar requires it[^1], keeping it in the production environment does
not cause errors, so it is not conditional on `DJANGO_DEBUG_TOOLBAR`.

[^1]: https://django-debug-toolbar.readthedocs.io/en/latest/installation.html