Branch `urlpatterns` based on `settings.DEBUG_TOOLBAR`
From the installation docs[^1],
django-debug-toolbar URLs must be added to `urlpatterns` for
the toolbar to be available.

Similar to the previous commit, to avoid breaking things in the
production environment, have the inclusion of the DDT urls
be dependent on the `DEBUG_TOOLBAR` setting,
which should never be `True` in an environment where
django-debug-toolbar is unavailable.

[^1]: https://django-debug-toolbar.readthedocs.io/en/latest/installation.html