Branch `settings.py` based on `DJANGO_DEBUG_TOOLBAR` env variable
We would like to make the django debug toolbar available for
local development.

From https://django-debug-toolbar.readthedocs.io/en/latest/installation.html,
three changes to `settings.py` are needed.
However, two of them (the additions to `INSTALLED_APPS` and `MIDDLEWARE`)
would cause an error if django-debug-toolbar is unavailable,
which is the case for our production environment.

Hence, make the above additions conditional on a boolean environment
variable named `DJANGO_DEBUG_TOOLBAR`.
If `DJANGO_DEBUG_TOOLBAR` is not set to 1 or true,
keep `INSTALLED_APPS` and `MIDDLEWARE` as-is.

The third change needed to enable the toolbar is the addition of a
`INTERNAL_IPS` setting.
Since keeping this setting does not cause errors in the production
environment, it is not conditional on `DJANGO_DEBUG_TOOLBAR`.