Branch `settings.py` based on `DJANGO_DEBUG_TOOLBAR` env variable
From django debug toolbar's installation docs[^1],
three changes to `settings.py` are needed to enable the toolbar.
However two of them (the additions to `INSTALLED_APPS` and `MIDDLEWARE`)
would cause an error if django-debug-toolbar is unavailable,
which is the case for our production environment.

Hence, make the above additions conditional on a boolean environment
variable named `DJANGO_DEBUG_TOOLBAR`.
If `DJANGO_DEBUG_TOOLBAR` does not evaluate to `True`,
keep `INSTALLED_APPS` and `MIDDLEWARE` as-is.

The third change needed to enable the toolbar is the addition of a
`INTERNAL_IPS` setting.
Since keeping this setting does not cause errors in the production
environment, it is not conditional on `DJANGO_DEBUG_TOOLBAR`.

Unfortunately it is not possible to test for these branches due to how
`pytest-django` sets up the test environment (which involves loading in
the settings prior to initialising individual tests).
The branches have been hence `pragma: no cover`ed.

[^1]: https://django-debug-toolbar.readthedocs.io/en/latest/installation.html