Add test for branching in `urls.py`
- Test that the `urlpatterns` in `urls.py` includes URL patterns from
`django-debug-toolbar` in debug mode, but not in production, as defined
by the `DEBUG` and `DEBUG_TOOLBAR` settings.

- Due to the nature of the tests, `actions.urls` must be imported after
the settings are patched, so we unfortunately violate PEP 8[^1].

- Python caches imported modules[^2] in `sys.modules`.
In order to fully isolate the tests, use a pytest fixture to remove
`actions.urls` from the module cache before and after a test that
imports it.
If this is not done, depending on the test execution order,
the cached `actions.urls` module (which may include the debug toolbar
URL patterns) could leak into other tests in the suite,
which assume that debug mode is off.

[^1]: https://peps.python.org/pep-0008/#imports
[^2]: https://docs.python.org/3.10/reference/import.html#the-module-cache