Add tests
This commit will be squashed into the previous commit.

- The branches in `settings.py` are tested by monkeypatching the
environment variable, and the branch in `urls.py` is tested by
monkeypatching the django settings.

- In both cases, the relevant module needs to be imported after the patching.
To ensure tests are isolated and do not interfere with each other,
create a factory fixture to ensure that the module cache does not
contain the module before the test imports it, nor does the module
imported during the test end up in the cache and pollute other tests.