Fix DDT urls test to be compatible with DDT v6
Django debug toolbar v6 safeguards against using its urls if the debug
toolbar is not in settings.INSTALLED_APPS by raising an error.

Previously we tested for the absence/presence of the debug toolbar
urls based on the values of settings.DEBUG and settings.DEBUG_TOOLBAR,
which we patch. This is no longer good enough as the presence of the
debug toolbar in settings.INSTALLED_APPS is now a hard requirement for
the DDT urls.

The test now checks for the absence/presence of the debug toolbar urls
based on the expected values of `settings.DEBUG`,
`settings.DEBUG_TOOLBAR` and `settings.INSTALLED_APPS` when the
`DJANGO_DEBUG_TOOLBAR` environment variable is set to "1" or "0".

Closes #481