Rewrite `upgrade` recipe
Previously `upgrade` could not call `_uv` because we needed to forbid
single-package upgrades occurring simulataneously with a change in the
(global) cutoff timestamp. This is no longer a concern, so instead of
writing `upgrade` as a shebang recipe just let it use `_uv`.

Also add a `package-date` argument to pass to `exclude-newer-package`.
We name the argument `package-date` to specify that the user only needs
to provide a date, not a `PACKAGE=DATE` pair.