Rewrite `update-dependencies` recipe
Check the existing package-specific cutoffs in the lockfile and retain
the ones that occur later than the specified new cutoff timestamp.

This is kept as a shebang recipe at the moment - although there is a lot
of repetition with the `_uv` recipe so before adopting this as it is,
it's worth reviewing whether this can be simplified by calling `_uv`.