Adjust `INSTALLED_APPS` in test for DDTv6 compatibility.
Django Debug Toolbar v6 requires that `debug_toolbar` is in `INSTALLED_APPS`
when using its URLs. So we need to add that to the test that uses it.

Remove unnecessary adjustment of `settings.DEBUG`.

Change to use `importlib.reload` over the pytest fixtures as this seems simpler.

Remove test parametrization because I think it made it harder to read these
very short tests and there isn't a meaningful amount of repeated code.

Group the tests by a class to make the names more less cumbersome and make it
easy to run these tests together if there are other tests added in future. Add
docstring.

Co-authored-by: Alice Wong <alice.wong@thedatalab.org>