Remove tests for `DJANGO_DEBUG_TOOLBAR` env var
Prompted by Slack discussion [^1] following a show and tell on these
tests, it seems best to remove them as they are imperfectly isolated
(details below) and a bit too complicated for the value they provide.

The Django debug toolbar is a dev dependency, so if we accidentally
deploy code that references it, we expect the build/deploy to fail
loudly, which seems sufficient for what these tests were trying to guard
against.

More detailed context:
The tests had to fiddle with Python's module cache in order to test
the presence or absence of the Django debug toolbar settings and URLs
based on the `DJANGO_DEBUG_TOOLBAR` env var. It turns out the
`remove_from_module_cache` fixture did not actually cater to all
possible states the module cache could be in [^2] prior to the test
running, so these unit tests were not truly isolated.
Here we just remove the fixture and the tests that used it instead of
further complicating the fixture.

[^1]: https://bennettoxford.slack.com/archives/C63UXGB8E/p1760004004418009
[^2]: https://bennettoxford.slack.com/archives/C63UXGB8E/p1760004435374159